added basic slide mechanic
I've ported over the basic slide movement from the Quake 1 Slide mod. I'm quite proud of this so far. As far as I can tell, it should control exactly the same as it does in the original quake slide mod. I've yet to add sound effects, or slide triggers, or most of the nice stuff, but the basic mechanic is there and is toggled by pressing F.